Refactor GraphQL schema and resolvers
Split the location.ts file into separate files (palletType.ts, storage.ts) for better modularity and organization
