+++ src/graphql/location.ts
@@ -20,26 +20,7 @@ export const typeDef = `
         locationType: String!
         storages: [Storage]!
     }
-
-    type Storage {
-        locationId: Int!
-        palletTypeId: Int!
-        amount: Int!
-        palletType: PalletType!
-    }
-
-    type PalletType {
-        palletTypeId: Int!
-        product: String
-        amount: Int!
-        storages: [Storage]!
-    }
-
-    type Mutation {
-        addAmountToStorage(locationId: Int!, amount: Int!): Storage
-        deleteAmountFromStorage(locationId: Int!, amount: Int!): Storage
-    }
-`;
+`
 
 export const resolvers = {
     Query: {
@@ -78,57 +59,8 @@ export const resolvers = {
                 throw new Error('Error retrieving all locations ')
             }
         },
-
-        allStorages:async () => {
-            try{
-                const allStorages = await Storage.findAll({include: PalletType})
-                return allStorages
-            } catch(error) {
-                console.log(error)
-                throw new Error('Error retrieving all storages')
-            }
-        },
-        allPalletTypes:async () => {
-            try{
-                const allPalletTypes = await PalletType.findAll()
-                return allPalletTypes
-            } catch(error) {
-                console.log(error)
-                throw new Error('error')
-            }
-        }
     },
     Mutation: {
-        addAmountToStorage: async (_: unknown, args: {locationId: number, amount: number}) => {
-            try{
-                const storage = await Storage.findByPk(args.locationId)
-                if (!storage) {
-                    throw new Error(`Storage with ID ${args.locationId} not found`)
-                }
 
-                
-                storage.amount += args.amount
-                await storage.save()
-
-                return storage;
-            } catch (error) {
-                console.error(error)
-                throw new Error(`Error adding amount to storage with ID ${args.locationId}`)
-            }
-            },
-            deleteAmountFromStorage: async (_: unknown, args: {locationId: number, amount: number}) => {
-                try{
-                    const storage = await Storage.findByPk(args.locationId)
-                    if (!storage) {
-                        throw new Error(`Storage with ID ${args.locationId} not found`)
-                    }
-                    storage.amount -= args.amount
-                    await storage.save()
-                    return storage
-                } catch (error){
-                    console.log(error)
-                    throw new Error(`Error deleting amount from storage with ID ${args.locationId}`)
-                }
-            },
         } 
     };
