+++ src/graphql/schema.ts
@@ -4,6 +4,8 @@ import { merge } from 'lodash' // to merge resolver strings easier
 import { typeDef as Location, resolvers as locationResolvers } from './location'
 import { typeDef as User, resolvers as userResolvers } from './user'
 import { typeDef as Order, resolvers as orderResolvers } from './order'
+import { typeDef as Storage, resolvers as storageResolvers } from './storage'
+import { typeDef as PalletType, resolvers as palletTypeResolvers } from './palletType'
 
 // this is extended, fake empty field added as you cannot have an empty type
 const Query = `
@@ -18,8 +20,7 @@ const Mutation = `
     addAmountToStorage(locationId: Int!, amount: Int!): Storage
     deleteAmountFromStorage(locationId: Int!, amount: Int!): Storage
   }
-`;
-
+`
 
 
 // resolvers from other files are merged to this
@@ -31,6 +32,6 @@ const resolvers = {
 
 // put everything together and make the schema
 export const schema = makeExecutableSchema({
-    typeDefs: [Query, Mutation, Location, User, Order],
-    resolvers: merge(resolvers, locationResolvers, userResolvers, orderResolvers),
+    typeDefs: [Query, Mutation, Location, User, Order, Storage, PalletType],
+    resolvers: merge(resolvers, locationResolvers, userResolvers, orderResolvers, storageResolvers, palletTypeResolvers),
 })
